Simplify packet inspector layout and toolbar
